:construction: Added My Fancy Animated Banner
Added in my experimental animated banner, although I did tweak the numbers a bit. Also – I decided to have the negative space be a div placed at the top of the site instead of create a positive space div below the site, mostly to make the code a little easier to manage.